+++ AdvertisementProject/service/src/main/java/com/senlainc/javacourses/petushokvaliantsin/service/advertisement/AdvertisementPhotoService.java
@@ -1,11 +1,13 @@
 package com.senlainc.javacourses.petushokvaliantsin.service.advertisement;
 
 import com.senlainc.javacourses.petushokvaliantsin.dto.advertisement.AdvertisementPhotoDto;
+import com.senlainc.javacourses.petushokvaliantsin.enumeration.EnumException;
 import com.senlainc.javacourses.petushokvaliantsin.enumeration.EnumLogger;
 import com.senlainc.javacourses.petushokvaliantsin.model.advertisement.AdvertisementPhoto;
 import com.senlainc.javacourses.petushokvaliantsin.repositoryapi.advertisement.IAdvertisementPhotoDao;
 import com.senlainc.javacourses.petushokvaliantsin.service.AbstractService;
 import com.senlainc.javacourses.petushokvaliantsin.serviceapi.advertisement.IAdvertisementPhotoService;
+import com.senlainc.javacourses.petushokvaliantsin.utility.exception.EntityNotExistException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,8 +29,7 @@ public class AdvertisementPhotoService extends AbstractService implements IAdver
     @Override
     @Transactional
     public boolean create(AdvertisementPhotoDto object) {
-        final AdvertisementPhoto advertisementPhoto = dtoMapper.map(object, AdvertisementPhoto.class);
-        advertisementPhotoDao.save(advertisementPhoto);
+        advertisementPhotoDao.save(dtoMapper.map(object, AdvertisementPhoto.class));
         LOGGER.info(EnumLogger.SUCCESSFUL_CREATE.getText());
         return true;
     }
@@ -36,6 +37,10 @@ public class AdvertisementPhotoService extends AbstractService implements IAdver
     @Override
     @Transactional
     public boolean delete(Long index) {
+        if (!advertisementPhotoDao.existsById(index)) {
+            throw new EntityNotExistException(String.format(EnumException.CLASS_WITH_FIELD_NOT_EXIST.getMessage(),
+                    AdvertisementPhoto.class.getSimpleName(), "index", index));
+        }
         advertisementPhotoDao.deleteById(index);
         LOGGER.info(EnumLogger.SUCCESSFUL_DELETE.getText());
         return true;
